requirement.txt
streamlit==1.37.1
pandas>=2.0.0
numpy>=1.24.0
azure-identity>=1.16.0
deltalake>=0.20.0
pyarrow>=14.0.0